Rename moonshot module to cdd (#586)
* rename moonshot module to cdd

* rename download_moonshot_data.py and test_moonshot.py
